clanetmock:renamed component
