+++ src/routes/clanetMock/advancedSettingsPanel/leverSettingsPanel/baseLever.svelte
