+++ src/routes/clanetMock/advancedSettingsPanel/leverSettingsPanel/eachLeverList/lever2.svelte
