+++ src/routes/clanetMock/advancedSettingsPanel/leverSettingsPanel/eachLeverList/lever3.svelte
