+++ src/routes/clanetMock/advancedSettingsPanel/leverSettingsPanel/eachLeverList/lever4.svelte
