+++ src/routes/clanetMock/advancedSettingsPanel/leverSettingsPanel/eachLeverList/lever5.svelte
