+++ src/routes/clanetMock/advancedSettingsPanel/pedalSettingsPanel/basePedal.svelte
