+++ src/routes/clanetMock/advancedSettingsPanel/pedalSettingsPanel/eachPedalList/pedalA.svelte
