+++ src/routes/clanetMock/advancedSettingsPanel/pedalSettingsPanel/eachPedalList/pedalB.svelte
