+++ src/routes/clanetMock/advancedSettingsPanel/pedalSettingsPanel/eachPedalList/pedalC.svelte
