+++ src/routes/clanetMock/mainControlPanel/baseControl.svelte
