+++ src/routes/clanetMock/mainControlPanel/eachSwitchPanel/aml.svelte
