+++ src/routes/clanetMock/mainControlPanel/eachSwitchPanel/idleVolume.svelte
