+++ src/routes/clanetMock/mainControlPanel/eachSwitchPanel/mKousoku.svelte
