+++ src/routes/clanetMock/mainControlPanel/eachSwitchPanel/makisugi.svelte
