+++ src/routes/clanetMock/mainControlPanel/eachSwitchPanel/sKousoku.svelte
