+++ src/routes/clanetMock/mainControlPanel/eachSwitchPanel/senkaiBrake.svelte
