+++ src/routes/clanetMock/mainControlPanel/eachSwitchPanel/senkaiFree.svelte
