+++ src/routes/clanetMock/mainControlPanel/eachSwitchPanel/stopSwitch.svelte
